requirements.txt
# 🌙 Moon Dev's Compliance Agent Requirements 🌙
# Created with ❤️ by Moon Dev's AI Assistant

# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
whisper>=1.1.10
openai>=1.51.0
anthropic>=0.5.0
groq>=0.4.0
# google-generativeai>=0.3.0  # Temporarily disabled due to protobuf conflict
# To re-enable Gemini, uncomment above and ensure compatible versions:
# google-generativeai==0.8.3
# protobuf==4.25.1  
# proto-plus==1.25.0
termcolor>=2.3.0
python-dotenv>=1.0.0
pathlib>=1.0.1

# For transcription
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# For image processing
pillow>=10.0.0

# For API access
requests>=2.31.0

# For data handling
pandas>=2.0.0

# For progress bars and arXiv paper downloader
# Updated by Moon Dev's AI Assistant on arXiv downloader project
tqdm>=4.66.0

backtesting>=0.3.3
ta-lib>=0.4.0
PyPDF2>=3.0.0
youtube-transcript-api>=0.6.2
pandas-ta>=0.3.14b0

# MetaTrader 5 Integration
MetaTrader5>=5.0.45

# Production Utilities
psutil>=5.9.0  # System monitoring
